move flexi-length
