+++ components/map/renderer/json.ts
@@ -1,4 +1,4 @@
-import { FlexiLength } from "@/components/map/renderer/flexi-length";
+import { FlexiLength } from "@/components/map/renderer/dimensions/flexi-length";
 import { FlexiPoint } from "@/components/map/renderer/flexi-point";
 import { nonNull } from "@dan-schel/js-utils";
 import { z } from "zod";
