+++ scripts/generate-map-geometry/lib/segment-instructions.ts
@@ -1,4 +1,4 @@
-import { FlexiLength } from "@/components/map/renderer/flexi-length";
+import { FlexiLength } from "@/components/map/renderer/dimensions/flexi-length";
 
 export type StraightSegmentInstruction = ReturnType<typeof straight>;
 export type CurveSegmentInstruction = ReturnType<typeof curve>;
