+++ scripts/generate-map-geometry/lib/utils.ts
@@ -1,5 +1,5 @@
 import { lineWidth } from "@/components/map/renderer/utils";
-import { flexi } from "@/components/map/renderer/flexi-length";
+import { flexi } from "@/components/map/renderer/dimensions/flexi-length";
 import { SegmentInstruction } from "@/scripts/generate-map-geometry/lib/segment-instructions";
 import { FlexiPoint } from "@/components/map/renderer/flexi-point";
 import { LineBuilder } from "@/scripts/generate-map-geometry/lib/line-builder";
