+++ scripts/generate-map-geometry/ptv/lines/frankston.ts
@@ -1,4 +1,4 @@
-import { flexi } from "@/components/map/renderer/flexi-length";
+import { flexi } from "@/components/map/renderer/dimensions/flexi-length";
 import { flindersStreetToRichmond } from "@/scripts/generate-map-geometry/ptv/segments/flinders-street-to-richmond";
 import {
   defaultRadius,
