+++ scripts/generate-map-geometry/ptv/utils-city-loop.ts
@@ -1,4 +1,4 @@
-import { flexi } from "@/components/map/renderer/flexi-length";
+import { flexi } from "@/components/map/renderer/dimensions/flexi-length";
 import { fp } from "@/components/map/renderer/flexi-point";
 import { lineGap } from "@/scripts/generate-map-geometry/ptv/utils";
