+++ scripts/generate-map-geometry/ptv/utils.ts
@@ -1,4 +1,7 @@
-import { flexi, FlexiLength } from "@/components/map/renderer/flexi-length";
+import {
+  flexi,
+  FlexiLength,
+} from "@/components/map/renderer/dimensions/flexi-length";
 
 export const lineGap = flexi(5);
 export const long45 = Math.cos(Math.PI / 4);
